backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic[email]
passlib[bcrypt]
python-jose
python-dotenv
